Add more type annotations
Application moved to application.py to avoid circular imports
